refactor: separate components
