+++ src/App.tsx
@@ -9,7 +9,7 @@ import ProtectedRoute from './components/util/ProtectedRoute';
 import PublicRoute from './components/util/PublicRoute';
 import { Provider } from 'react-redux';
 import store from './store';
-import ChatPage from './components/routes/Chat';
+import ChatPage from './components/routes/chat/ChatsPage';
 
 function App() {
 	return (
