+++ src/components/routes/chat/ChannelsPanel.tsx
@@ -1,13 +1,8 @@
 import React from 'react';
-
-import Typography from '@material-ui/core/Typography';
-import Container from '@material-ui/core/Container';
 import { makeStyles } from '@material-ui/core/styles';
-import Page from '../Page';
 import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemText from '@material-ui/core/ListItemText';
-import { Paper, AppBar, Tabs, Tab, Avatar, ListItemAvatar } from '@material-ui/core';
+import { Paper, AppBar, Tabs, Tab } from '@material-ui/core';
+import ChannelListItem from './ChannelListItem';
 
 const dummyUsers = [
 	{ name: 'Barack Obama', src: 'https://www.biography.com/.image/t_share/MTE4MDAzNDEwNzg5ODI4MTEw/barack-obama-12782369-1-402.jpg' },
@@ -23,13 +18,6 @@ const dummyEvents = [
 ];
 
 const useStyles = makeStyles(theme => ({
-	mainContent: {
-		padding: theme.spacing(6, 2, 14, 2),
-		textAlign: 'center'
-	},
-	chatsRoot: {
-		display: 'flex'
-	},
 	channelsPanel: {
 		width: theme.spacing(40)
 	},
@@ -37,22 +25,14 @@ const useStyles = makeStyles(theme => ({
 		minHeight: '80vh',
 		maxHeight: '80vh',
 		overflow: 'auto'
-	},
-	flexGrow: {
-		flexGrow: 1
 	}
 }));
 
-function UserListItem({ name, src }: {name: string; src: string}) {
-	return <ListItem button>
-		<ListItemAvatar>
-			<Avatar alt={name} src={src}/>
-		</ListItemAvatar>
-		<ListItemText primary={name} />
-	</ListItem>;
+interface ChannelsPanelProps {
+	onChannelSelected: (channel: string) => void;
 }
 
-export function ChannelsPanel() {
+export default function ChannelsPanel({ onChannelSelected }: ChannelsPanelProps) {
 	const classes = useStyles();
 
 	const [chatPanelValue, setChatPanelValue] = React.useState(0);
@@ -67,26 +47,9 @@ export function ChannelsPanel() {
 					<Tab label="Events" />
 				</Tabs>
 			</AppBar>
-			<List component="nav" aria-label="main mailbox folders" className={classes.channelsList} >
-				{channelList.map(user => <UserListItem {...user} />)}
+			<List component="nav" aria-label="channels" className={classes.channelsList} >
+				{channelList.map((channel, index) => <ChannelListItem key={index} {...channel} onClick={() => onChannelSelected(channel.name)} />)}
 			</List>
 		</div>
 	</Paper>;
-}
-
-export default function ChatPage() {
-	const classes = useStyles();
-
-	return (
-		<Page>
-			<Container maxWidth="xl" component="main" className={classes.mainContent}>
-				<Paper elevation={3} className={classes.chatsRoot}>
-					<ChannelsPanel />
-					<div className={classes.flexGrow}>
-						<Typography>hi</Typography>
-					</div>
-				</Paper>
-			</Container>
-		</Page>
-	);
 }
